+++ test/unit/fuzz-tools.test.js
@@ -328,6 +328,7 @@ import * as strasbergBdiV348 from '../../lib/strasberg-bdi-v348.js';
 import * as fazekasV349 from '../../lib/fazekas-v349.js';
 import * as tscherneClosedV350 from '../../lib/tscherne-closed-v350.js';
 import * as goligherHemorrhoidsV351 from '../../lib/goligher-hemorrhoids-v351.js';
+import * as lanskyV352 from '../../lib/lansky-v352.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -601,6 +602,7 @@ const MODULES = {
   'fazekas-v349.js': fazekasV349,
   'tscherne-closed-v350.js': tscherneClosedV350,
   'goligher-hemorrhoids-v351.js': goligherHemorrhoidsV351,
+  'lansky-v352.js': lanskyV352,
   'tb-testing.js': tbTesting,
 };